+++ node/clients/catalog.ts
@@ -1,4 +1,5 @@
-import { InstanceOptions, IOContext, ExternalClient} from '@vtex/api'
+import type { InstanceOptions, IOContext } from '@vtex/api'
+import { ExternalClient } from '@vtex/api'
 
 export class Catalog extends ExternalClient {
   constructor(context: IOContext, options?: InstanceOptions) {