+++ node/globals.ts
@@ -1,10 +1,10 @@
-import {
+import type {
   IOContext,
-  MetricsAccumulator,
   ParamsContext,
   RecorderState,
   ServiceContext,
 } from '@vtex/api'
+import { MetricsAccumulator } from '@vtex/api'
 
 import type { Clients } from './clients'